Add files via upload
The Accident Alert IoT project utilizes Arduino, GPS, and GSM modules to enhance road safety. The system involves a GPS module for real-time location tracking, an accelerometer for accident detection, and a GSM module for communication. The Arduino processes data, and if an accident is detected, it sends emergency alerts with GPS coordinates to predefined contacts. The project aims to provide a quick and automated response in critical situations, contributing to improved road safety by reducing response times and providing accurate location information. Including another Arduino module that detects the speed  of a vehicle that travels between two IR sensors, in which alarm buzzes if the vehicle exceeds the normal speed and displays each vehicles speed using a LCD display.